% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_sde.R
\name{phylo_sde_0}
\alias{phylo_sde_0}
\title{phylo_sde}
\usage{
phylo_sde_0(tr, rt_value, N, theta, model, method, fossils = NULL, ...)
}
\arguments{
\item{tr}{An object of class \code{phylo} from the ape package.  In this version the CIR
process parameters alpha, mu, and sigma for each branch are within vectors in the same order
as the edge (branch) labelling.}

\item{rt_value}{Value at the root of \code{tr}.}

\item{N}{Data imputation frequency.}

\item{theta}{Matrix of parameter values for each edge of the tree.}

\item{model}{A list containing drift, diffusion, and the partial differentiation of diffusion as quoted
expressions using method quote. For the Euler scheme
the drift coefficient as \code{drift}, the diffusion coefficient as
\code{diffusion}, and the partial differentiation of
\code{diffusion} by \code{x} as \code{dx_diffusion} is required.
See the Examples.}

\item{method}{Specified as either "euler" or "milstein."}

\item{...}{Not used.}
}
\value{
A list of time series projects for each simulated path equal to the length of
the number of branches in the \code{tr} object.
}
\description{
Simulate a Cox-Ingersoll-Ross Diffusion Process in the Tree of Life
}
