% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_logL.R
\name{tree_logL}
\alias{tree_logL}
\title{tree_logL}
\usage{
tree_logL(tr, tipdata, lst, alpha, mu, sigma, model, method, fossils = NULL,
  ...)
}
\arguments{
\item{tr}{An object of class \code{phylo} from the ape package.  In this version the CIR
process parameters alpha, mu, and sigma for each branch are within vectors in the same order
as the edge (branch) labelling.}

\item{tipdata}{A numeric vector containing tip values in the same order as the tip labels in \code{tr$tip.label}.}

\item{lst}{A list of time series projects for each simulated path equal to the length of
the number of branches in the \code{tr} object.}

\item{alpha}{Set to NULL if alpha is to be estimated,
otherwise set to a numeric value or a numeric vector specifying
the value of the parameter for all the branches/edges.
In the latter case, the values must be specifying in the same order
as the edges in the \code{tr} object.}

\item{mu}{As \code{alpha}.}

\item{sigma}{As \code{alpha}.}

\item{model}{A list containing drift, diffusion, and the partial differentiation of diffusion as quoted
expressions using method quote. For the Euler scheme
the drift coefficient as \code{drift}, the diffusion coefficient as
\code{diffusion}, and the partial differentiation of
\code{diffusion} by \code{x} as \code{dx_diffusion} is required.}

\item{method}{Specified as either "euler" or "milstein."}

\item{fossils}{A numeric vector containing the tip values for every fossil added to the tree.}

\item{...}{Not used.}
}
\value{
logL, an integer.
}
\description{
Calculate the loglikelihood of a Cox-Ingersoll-Ross Diffusion Process in the Tree of Life
}
\details{
This function caluclates the log likelihood.
}
\seealso{
\code{\link{fossil_id}}
}
